refactor: Add coverUrl and headshotUrl fields to Movie and Actor models
This commit adds the coverUrl field to the Movie model and the headshotUrl field to the Actor model in the schema.prisma file. These fields will store the URLs of the cover image for movies and the headshot image for actors. This change allows for displaying images associated with movies and actors in the application, enhancing the visual experience for users.